Change CMakeLists to explicitly look for Qt4
In some cases/distros there are 3 cmake modules for Qt:
  - FindQt3.cmake
  - FindQt4.cmake
  - FindQt.cmake

By doing find_package(Qt 4.7), we were using the third rule. CMake will complain
in systems with both Qt3 and Qt4 installed and in PATH, asking for an extra argument
to define which Qt version should be used.

As we completely removed the dependency on Qt3, it is safe to modify CMakeLists
to find_package(Qt4 4.7), so we explicitly look for Qt4 >= 4.7.

Thanks to Mailson Menezes for reporting this issue.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@openbossa.org>